+++ tests/testthat/test-cTMed-boot-med.R
@@ -1,4 +1,4 @@
-## ---- test-cTMed-delta-med
+## ---- test-cTMed-boot-med
 lapply(
   X = 1,
   FUN = function(i,
@@ -6,7 +6,7 @@ lapply(
                  tol) {
     message(text)
     testthat::test_that(
-      paste(text, "DeltaMed"),
+      paste(text, "BootMed"),
       {
         testthat::skip_on_cran()
         total <- 0.0799008
@@ -58,10 +58,18 @@ lapply(
           ),
           nrow = 9
         )
-        delta <- DeltaMed(
+        delta_t <- 2
+        R <- 1000
+        mc <- MCPhi(
           phi = phi,
           vcov_phi_vec = vcov_phi_vec,
-          delta_t = 2,
+          R = R,
+          seed = 42
+        )$output
+        boot <- BootMed(
+          phi = mc,
+          phi_hat = phi,
+          delta_t = delta_t,
           from = "x",
           to = "y",
           med = "m"
@@ -69,7 +77,7 @@ lapply(
         testthat::expect_true(
           all(
             (
-              answer - summary(delta)$est
+              answer - summary(boot)$est
             ) <= tol
           )
         )
@@ -128,35 +136,50 @@ lapply(
           ),
           nrow = 9
         )
-        delta <- DeltaMed(
+        delta_t <- 2
+        R <- 1000
+        mc <- MCPhi(
           phi = phi,
           vcov_phi_vec = vcov_phi_vec,
+          R = R,
+          seed = 42
+        )$output
+        boot <- BootMed(
+          phi = mc,
+          phi_hat = phi,
           delta_t = 1:5,
           from = "x",
           to = "y",
           med = "m"
         )
-        print(delta)
-        summary(delta)
-        confint(delta, level = 0.95)
-        plot(delta)
-        delta <- DeltaMed(
-          phi = phi,
-          vcov_phi_vec = vcov_phi_vec,
+        print(boot)
+        summary(boot)
+        confint(boot)
+        plot(boot)
+        print(boot, type = "bc")
+        summary(boot, type = "bc")
+        confint(boot, type = "bc")
+        plot(boot, type = "bc")
+        boot <- BootMed(
+          phi = mc,
+          phi_hat = phi,
           delta_t = 1,
           from = "x",
           to = "y",
           med = "m"
         )
-        print(delta)
-        summary(delta)
-        confint(delta, level = 0.95)
+        print(boot)
+        summary(boot)
+        confint(boot, level = 0.95)
+        print(boot, type = "bc")
+        summary(boot, type = "bc")
+        confint(boot, type = "bc", level = 0.95)
         testthat::expect_error(
-          plot(delta)
+          plot(boot)
         )
       }
     )
   },
-  text = "test-cTMed-delta-med",
+  text = "test-cTMed-boot-med",
   tol = 0.01
 )